// front-end/src/components/main-grid/types/main-view.ts
export interface MainView {
  id: number;
  tb: string;
  period: string;
  model: string;
  cartridgeType: string;
  type: string;
  ip: string;
  vsp: string;
  uid: string;
  address: string;
  lastSeen: string;
  location: string;
  serial: string;
  mileadge: string;
  points: string;
  refreshRequest: string;
  refreshDate: Date;
  refreshStatus: string;
  refreshSolution: string;
  newSerial: string;
  comment: string;
  keywords: string;
  grouping: string;

  repairRequest: string;
  repairDate: Date;
  repairStatus: string;
  repairSolution: string;
  path : string
}